LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-tables/table-model-fixup-2-expected.txt

Table fixup
Check that replaced elements display types are handled properly in fixup

Both input fields should look identical, on their own line:

Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:

Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)


Both images should look identical, on their own line:

Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)



Both images should look identical and be on the same line:

Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:

Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:

Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)


Both images should look identical, on their own line:

Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)



Both images should look identical and be on the same line:

Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top)


Both input fields should look identical, on their own line:

Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)



Both input fields should look identical and be on the same line:

Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)


Both images should look identical, on their own line:

Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)



Both images should look identical and be on the same line:

Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top)


Both text inputs should share the same line:

Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be inline-table and are considered inline -- img elements


Both text inputs should stand on their own line:

Replaced elements outside a table cannot be table and are considered block -- input=text elements


Both button inputs should stand on their own line:

Replaced elements outside a table cannot be table and are considered block -- input=button elements


Both file inputs should stand on their own line:

Replaced elements outside a table cannot be table and are considered block -- input=file elements


Both images should stand on their own line:

Replaced elements outside a table cannot be table and are considered block -- img elements


Both text inputs should share the same line:

Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be table-row and are considered inline -- img elements


Both text inputs should share the same line:

Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements


Both text inputs should share the same line:

Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be table-column and are considered inline -- img elements


Both text inputs should share the same line:

Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be table-cell and are considered inline -- img elements


Both text inputs should share the same line:

Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements


Both button inputs should share the same line:

Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements


Both file inputs should share the same line:

Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements


Both images should share the same line:

Replaced elements outside a table cannot be table-caption and are considered inline -- img elements



PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-row and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-row and are considered inline -- img elements (top) assert_equals: expected 817.875 but got 849.875
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-column and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-column and are considered inline -- img elements (top) assert_equals: expected 1579.875 but got 1611.875
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (width)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- input elements (top)
PASS Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (width)
FAIL Replaced elements inside a table cannot be table-cell and are considered inline -- img elements (top) assert_equals: expected 2331.875 but got 2363.875
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=button elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- input=file elements
PASS Replaced elements outside a table cannot be inline-table and are considered inline -- img elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=text elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=button elements
PASS Replaced elements outside a table cannot be table and are considered block -- input=file elements
PASS Replaced elements outside a table cannot be table and are considered block -- img elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- input=file elements assert_equals: expected 3846.875 but got 3873.875
FAIL Replaced elements outside a table cannot be table-row and are considered inline -- img elements assert_equals: expected 3996.875 but got 4028.875
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-row-group and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- input=file elements assert_equals: expected 4399.875 but got 4426.875
FAIL Replaced elements outside a table cannot be table-row-group and are considered inline -- img elements assert_equals: expected 4549.875 but got 4581.875
PASS Replaced elements outside a table cannot be table-column and are considered inline inline -- input=text elements
PASS Replaced elements outside a table cannot be table-column and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- input=file elements assert_equals: expected 4952.875 but got 4979.875
FAIL Replaced elements outside a table cannot be table-column and are considered inline -- img elements assert_equals: expected 5102.875 but got 5134.875
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-cell and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- input=file elements assert_equals: expected 5505.875 but got 5532.875
FAIL Replaced elements outside a table cannot be table-cell and are considered inline -- img elements assert_equals: expected 5655.875 but got 5687.875
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=text elements
PASS Replaced elements outside a table cannot be table-caption and are considered inline -- input=button elements
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- input=file elements assert_equals: expected 6058.875 but got 6085.875
FAIL Replaced elements outside a table cannot be table-caption and are considered inline -- img elements assert_equals: expected 6208.875 but got 6240.875
